+++ src/netcdf_writer_coordinates.jl
@@ -1,3 +1,6 @@
+import ..Interpolators: PressureInterpolator, era5_pressure_levels
+import ClimaCore: Remapping
+
 """
     AbstractZSamplingMethod
 
@@ -32,6 +35,81 @@ p ~ p₀ exp(-z/H)
 H is assumed to be 7000 m, which is a good scale height for the Earth atmosphere.
 """
 struct FakePressureLevelsMethod <: AbstractZSamplingMethod end
+
+"""
+    RealPressureLevelsMethod
+
+Interpolate exactly to the points of the space whose vertical coordinate consists
+of `PPoint`s.
+
+This should be reused across all diagnostics that will be written in pressure
+coordinates.
+
+!!! note "Compatible writers"
+    This is only compatible with the `NetCDFWriter`.
+
+!!! note "Same space as pressure field"
+    The pressure field computed by `compute_pfull!` must be on the same space
+    as interpolated field. If this is not the case, you may need to interpolate
+    the field defined on a face space to a field defined on a center space or
+    vice versa.
+"""
+struct RealPressureLevelsMethod{PFULL_INTP, MAPPING} <: AbstractZSamplingMethod
+    pfull_intp::PFULL_INTP
+    diag_to_scratch::MAPPING
+    pressure_attribs::NamedTuple
+end
+
+"""
+    RealPressureLevelsMethod(
+        pfull_field,
+        t;
+        pressure_attribs = (;),
+        pressure_levels = era5_pressure_levels(),
+        pressure_intp_kwargs = (; ),
+    )
+
+Sample points defined on the vertical coordinate of a space that contains
+`PPoint`s.
+
+The keyword arguments `pressure_levels` and `pressure_intp_kwargs` passed to
+`RealPressureLevelsMethod` are passed to the constructor of
+`ClimaCore.Remapping.PressureInterpolator`.
+
+The default attributes are
+- long\\_name: "pressure",
+- units: "Pa",
+- stored\\_direction: "increasing",
+- standard\\_name: "air\\_pressure".
+"""
+function RealPressureLevelsMethod(
+    pfull_field,
+    t;
+    pressure_attribs = (;),
+    pressure_levels = era5_pressure_levels(),
+    pressure_intp_kwargs = (;),
+)
+    pfull_intp = PressureInterpolator(
+        pfull_field,
+        t;
+        pressure_levels,
+        pressure_intp_kwargs,
+    )
+    return RealPressureLevelsMethod(
+        pfull_intp,
+        Dict{ScheduledDiagnostic, Fields.Field}(),
+        pressure_attribs,
+    )
+end
+
+"""
+    pressure_space(pfull_intp::PressureInterpolator)
+
+Return the space where the points along the vertical are `PPoint`s.
+"""
+function pressure_space(pressure_sampling_method::RealPressureLevelsMethod)
+    Remapping.pressure_space(pressure_sampling_method.pfull_intp.pressure_intp)
+end
 
 """
     add_dimension!(nc::NCDatasets.NCDataset,
@@ -258,11 +336,22 @@ function target_coordinates(
     return Array(parent(Fields.coordinate_field(cspace).z))[:, 1]
 end
 
+function target_coordinates(
+    space::S,
+    num_points,
+    ::RealPressureLevelsMethod,
+) where {
+    S <:
+    Union{Spaces.CenterFiniteDifferenceSpace, Spaces.FaceFiniteDifferenceSpace},
+}
+    return Array(parent(Fields.coordinate_field(space).p))[:, 1]
+end
+
 # Column
 function add_space_coordinates_maybe!(
     nc::NCDatasets.NCDataset,
     ::Spaces.FiniteDifferenceSpace,
-    num_points_z,
+    num_points,
     hpts,
     zpts::Vector{FT};
     z_sampling_method,
@@ -270,6 +359,28 @@ function add_space_coordinates_maybe!(
     interpolated_physical_z = nothing, # Not needed here, but needed for consistency of
     # interface and dispatch
 ) where {FT <: AbstractFloat}
+    # When the writer is initialized from a 3D space but we're writing a 1D column field,
+    # num_points will be a multi-element tuple (e.g., (num_lon, num_lat, num_z)).
+    # We need to extract only the vertical dimension (the last element).
+    num_points_z = length(num_points) > 1 ? (last(num_points),) : num_points
+    if z_sampling_method isa RealPressureLevelsMethod
+        name = "pressure_level"
+        pfull_dimension_exists = dimension_exists(nc, name, num_points_z)
+
+        if !pfull_dimension_exists
+            add_dimension!(
+                nc,
+                name,
+                zpts;
+                long_name = "pressure",
+                units = "Pa",
+                stored_direction = "increasing",
+                standard_name = "air_pressure",
+                z_sampling_method.pressure_attribs...,
+            )
+        end
+        return [name]
+    end
     name, _... = names
     z_dimension_exists = dimension_exists(nc, name, num_points_z)